backend/requirements.txt
steam-web-api>=1.0.0
flask>=2.0.0
flask-cors>=3.0.0
python-dotenv>=1.0.0
